+++ CHANGES.md
@@ -1,5 +1,14 @@
 CHANGES
 =======
+Version 1.1.0 - 2026-08-XX
+------------------
+- Black hole
+    [X] Appears when an event is received
+    [X] Disappears after a time elapsed
+    [ ] Moves through the map
+    [ ] Affects other players
+    [ ] Kills players on contact
+
 Version 1.0.1 - 2026-08-04
 ------------------
 - Black hole animation only